// src/model/VersionModel.ts
/**
 * Created by guoyan on 2018/4/3.
 */
export class VersionModel {


  versionId: string;
  branchId: string;
  branchName: string;
  projectName: string;
  platform: string;
  createTime: string;
  packageName: string;
  branchDescription: string;
  creator: string;
  newVersionId: string;
  plistFilePath: string;
  iconPath: string;
  downloadNum: string;
  ewmImgPath: string;   // 二维码图片的地址
  filePath: string;   // 安装包的下载地址
  versionCode: string;
  versionName: string;
  fileSize: string;
  versionDescription: string;
}